[MODIFY] main 일부 entity들을 posts 모듈로 이동한 것에 의한 entity 삭제와 생성
