--- DMtime/main/src/main/java/com/jaemin/main/domain/entity/Posts.kt
@@ -1,5 +1,0 @@
-package com.jaemin.main.domain.entity
-
-data class Posts(
-    val numberOfPages : Int,
-    val posts : List<Post>)
