--- DMtime/main/src/main/java/com/jaemin/main/domain/entity/Uploader.kt
@@ -1,5 +1,0 @@
-package com.jaemin.main.domain.entity
-
-import com.google.gson.annotations.SerializedName
-
-data class Uploader(val username : String)
