+++ DMtime/posts/src/main/java/com/jaemin/posts/domain/entity/PatchedPost.kt
@@ -1,0 +1,8 @@
+package com.jaemin.posts.domain.entity
+
+
+data class PatchedPost(
+    val imageIds: List<Int>,
+    val content: String,
+    val title: String
+)
