+++ DMtime/posts/src/main/java/com/jaemin/posts/domain/entity/PostPreview.kt
@@ -1,0 +1,13 @@
+package com.jaemin.posts.domain.entity
+
+
+data class PostPreview(
+    val id: Int,
+    val likes: Int,
+    val numberOfComments: Int,
+    val postedDatetime: String,
+    val title: String,
+    val uploader: Uploader,
+    val views: Int,
+    val whetherExistImage: Boolean
+)
