+++ DMtime/posts/src/main/java/com/jaemin/posts/domain/entity/Posts.kt
@@ -1,0 +1,6 @@
+package com.jaemin.posts.domain.entity
+
+data class Posts(
+    val numberOfPages: Int,
+    val posts: List<PostPreview>
+)
