+++ DMtime/posts/src/main/java/com/jaemin/posts/domain/entity/Uploader.kt
@@ -1,0 +1,3 @@
+package com.jaemin.posts.domain.entity
+
+data class Uploader(val username: String)
